Refactor code to avoid post processing
- Remove postprocessing in favour of per tag processing.
- Set only required attributes to True instead of setting all to True/False
- Set run formatting to font attribute to introduce more styles
  in the future versions. The `sub` and `sup` styles are set to font only
- Add paragraph formatting attributes to style them in
  the future versions
- Fix "br" tag processing. Insert a run break instead of starting a new
  paragraph.
- Extend "pre" tag test to include whitespace testing

The text portions are added to the document in `handle_data`
instead of in `handle_endtag` for block level tags.
The code used to defer adding of text runs and attributes till
the time when the closing block tag was available. This change does
not store these items of data. Rather it adds text data and
attributes as they are available. This is more typical for parser
based approach (python HTMLParser class) than for DOM based
approach. In the latter case the data is stored and the processing
is deferred until later stage.