Resize images to fit within the page usable area
To specify an image size, use the `width` and `height` attributes on the
`<img>` element. Rich text editors like TinyMCE use these attribute to
control the image size.
Though the HTML attributes are unit-less, interpret them as pixels to
allow conversion between attribute values and inches, the latter being
used as measurement unit in docx document.

When images go beyond the writable text area, resize them to fit within
the page. The usable area was determined by resizing an image with
LibreOffice in the default template document. The maximum height for an
image in a document is capped to 8.1 inches.

Without DPI information in most images, the default value of 72 DPI is
used.
https://python-docx.readthedocs.io/en/latest/api/document.html#docx.document.Document.add_picture

Reviewed-by: Jon Dufresne
Reviewed-by: Roman Danilov