Use default resolution when scaling image size
Scaling the images based on their resolution is good when the image size
is unconstrained, it prevents images with a high resolution from
appearing larger than images with a lower resolution but the same pixel
size.

When the image size is constrained, ignore resolution and scale image to
the constrained size.

The default DPI is now used in the project, moved the constant from the
tests to html2docx/image.py and renamed it to `DEFAULT_DPI` to better
describe it.

Reviewed-by: Jordan Arentsen